+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `configauth`: remove interceptor functions from the ServerAuthenticator interface (#4583)
 - Replace ConfigMapProvider and ConfigUnmarshaler in collector settings by one simpler ConfigProvider (#4590)
 - Remove deprecated consumererror.Combine (#4597)
+- Remove `configmapprovider.NewDefault`, `configmapprovider.NewExpand`, `configmapprovider.NewMerge` (#4600)
 
 ## 💡 Enhancements 💡